Created the simulation file.
The data structure msg in the udp-sender.c file contains a data stucture called msg too. In this data structure there is an array uint16_t type of ten elements. They can be set in order to the sensed values by the sensor. Right now the values are hard coded in order to see how it works the source code.