The power power consumption data are avalaible and ready to compute.
Before to activate the sensor and after to deactivate it, it has been
put a timestamp in order to see how much time the CO2, CO and temp
sensors take to make the measures. In the simulation enviroinment Cooja
this time is equal to zero. I think is due to the simulation
enviroinment itself since that Cooja does not consider this kind of
characteristic of pollution sensors. Starting to implement the sink side
in order to parse and compute the received data.