Duty cycle available, energy consumption almost.
d timing data to compute the power consumption are correctly extracted
and set in the respective data structure. After that the duty cycle is
printed out (a matemathical trick is used in order to print out the
"floating" value). The energy consumption of the listening is printed
out. Right now the energy consumption below 1 mJ is considered zero for
not-floating-type supported by Contiki reason.