All energy consumption are avalaible.
according to the datasheet) to 4 seconds which is an arbitrary values
chosen for testing reasons. The energy consumption of CO, CO2 and temp
sensor are obtained and printed out.
The project NES situation is the following: all energy consumptions are
now available. The next step is to compute the energy consumption
percentage of each module in each sensor.